# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Podcast Platform API",
        "description": "API documentation for the Podcast Platform",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/podcasts": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get all podcasts",
                "operationId": "74471147bb5a65d449b5357445399019",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "rating",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Create a new podcast",
                "operationId": "b653b8c06d3bde71f8996de1f4b426ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PodcastRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Podcast created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/podcasts/category/{categoryId}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get podcasts by category",
                "operationId": "1a894b3e4669f1c23259e196d12a3bfd",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/podcasts/search": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Search podcasts",
                "operationId": "5469c2bf94ab12401a6f4ffface1fb67",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/podcasts/{id}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get podcast by ID",
                "operationId": "e313077915b72667e295fc651f1aa955",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Update a podcast",
                "operationId": "9cf5198706fcc4e2630ab52034048de9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PodcastRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Podcast updated successfully"
                    },
                    "404": {
                        "description": "Podcast not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Delete a podcast",
                "operationId": "ad34860ba7e2917accdf74ad7558b605",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast deleted successfully"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PodcastRequest": {
                "title": "Podcast Request",
                "required": [
                    "title",
                    "slug",
                    "description",
                    "category_id",
                    "author"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Tech Talk"
                    },
                    "slug": {
                        "type": "string",
                        "example": "tech-talk"
                    },
                    "description": {
                        "type": "string",
                        "example": "A podcast about technology"
                    },
                    "image": {
                        "type": "string",
                        "example": "podcasts/tech-talk.jpg"
                    },
                    "category_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "rating": {
                        "type": "integer",
                        "example": 4
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}